requirements.txt
Flask>=2.3.0
mutagen>=1.47.0
Pillow>=10.1.0
musicbrainzngs>=0.7.1
werkzeug>=2.3.0
flask-cors>=4.0.0
python-magic>=0.4.27
requests>=2.31.0
pytest>=7.4.0
pytest-flask>=1.2.0
